gnirsSort: Write all.list for Tellurics
checkdata:
If the exposure times are not all the same then try to guess which is correct.  This initial version just looks at the exposure time frequency and uses the files with the most frequent exposure time.  If there are equal numbers of files with multiple exposure times it gives an error and exits.  In the future we could factor in the order that the data were collected, making the assumption that the final exposure time is the correct one.  The file list is updated and the original is saved as <list>.bak.